+++ openqaoa/openqaoa-azure/backends/__init__.py
@@ -1,0 +1,1 @@
+from .devices import DeviceAzure
