+++ openqaoa/openqaoa-braket/backends/__init__.py
@@ -1,0 +1,2 @@
+from .devices import DeviceAWS
+from .qaoa_braket_qpu import QAOAAWSQPUBackend
